requirements.txt
coverage==3.5.1
mock==0.7.0
django-nose==1.0
nose==1.1.2
pyzen==0.3.2